+++ kwcoco/__init__.py
@@ -269,7 +269,7 @@ Testing:
 
 """
 
-__version__ = '0.7.8'
+__version__ = '0.8.0'
 
 
 __submodules__ = {